chore: move packing method to pack3d package
