+++ pack3d/pack3d.go
@@ -1,0 +1,475 @@
+package pack3d
+
+import (
+	"encoding/json"
+	"fmt"
+	"log"
+	"math"
+	"os"
+	"time"
+
+	"github.com/fogleman/fauxgl"
+)
+
+// Applies rotation, scaling, and co-packing if specified.
+func Pack(input, output string) {
+
+	// Load config
+	var config Config
+	file, err := os.ReadFile(input)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	err = json.Unmarshal([]byte(file), &config)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	type TransMap struct {
+		Filename          string
+		Transformation    [4][4]float64
+		VolumeWithSpacing float64
+	}
+
+	var (
+		singleStlSize  []fauxgl.Vector
+		scaleStl       []fauxgl.Matrix
+		mfgRotationStl []fauxgl.Matrix
+		done           func()
+		totalVolume    float64
+		iterations     int
+		srcStlNames    []string
+		transMaps      []TransMap
+	)
+
+	model := NewModel()
+	scale := 1.0
+	var scaleMatrix fauxgl.Matrix
+	var mfgRotationMatrix fauxgl.Matrix
+	var ok bool
+
+	spacing := config.Spacing / 2.0
+
+	// frameSize is the vertex in the first quadrant
+	frameSize := fauxgl.V(config.BuildVolume[0], config.BuildVolume[1], config.BuildVolume[2])
+	buildVolume := config.BuildVolume[0] * config.BuildVolume[1] * config.BuildVolume[2]
+	//fmt.Println(frameSize)
+
+	/* Loading stl models */
+
+	// Tech Debt: there's unnecessary repetition of code inside the if-statement below.
+
+	coPackMap := make(map[string][]*Copack) // variable that contains co-packed mesh's data.
+	for _, item := range config.ConfigItems {
+
+		var mesh *fauxgl.Mesh
+		var err error
+
+		if item.Copack == nil {
+
+			// 1. load the mesh.
+			done = timed(fmt.Sprintf("loading mesh %s", item.Filename))
+			mesh, err = fauxgl.LoadMesh(item.Filename)
+			if err != nil {
+				panic(err)
+			}
+			done()
+
+			// 2. mesh centering.
+			mesh.Center()
+
+			// 3. apply the scaling to the mesh.
+			//    Notice that if scaling is to be applied, it is done
+			//    before the computation of the BoundingBox and volume.
+			scale = item.Scale
+			scaleMatrix = fauxgl.Scale(fauxgl.V(scale, scale, scale))
+			if scale != 1.0 {
+				done = timed("scaling mesh")
+				mesh.Transform(scaleMatrix)
+				done()
+			}
+
+			// 4. apply the manufacturing rotation mesh.
+			//    Notice that this is done before the computation of the BoundingBox and volume.
+			// IMPORTANT: do not confuse manufacturing orientation with the packing
+			//            orientations from the orientations provided by the annealing further on.
+			mfgRotationMatrix = getManufacturingOrientation(item)
+			mesh.Transform(mfgRotationMatrix)
+
+			// 5. update all the copies mesh for the json output.
+			size := mesh.BoundingBox().Size()
+			for i := 0; i < item.Count; i++ {
+				singleStlSize = append(singleStlSize, size)
+				srcStlNames = append(srcStlNames, item.Filename)
+				scaleStl = append(scaleStl, scaleMatrix)
+				mfgRotationStl = append(mfgRotationStl, mfgRotationMatrix)
+			}
+
+			fmt.Printf("  %d triangles\n", len(mesh.Triangles))
+			fmt.Printf("  %g x %g x %g\n", size.X, size.Y, size.Z)
+
+			// 6. coarse approx of its volume.
+			totalVolume += mesh.BoundingBox().Volume()
+
+		} else {
+
+			coPackMap[item.Filename] = item.Copack
+
+			// 1a. load the main co-packing mesh (the "parent" co-packing mesh, so to say).
+			done = timed(fmt.Sprintf("loading the main co-packing mesh %s", item.Filename))
+			mesh, err = fauxgl.LoadMesh(item.Filename)
+			if err != nil {
+				panic(err)
+			}
+			done()
+
+			// 1b. load the co-packed meshes (the "children" of the "parent" co-packing mesh, so to say).
+			for _, cp := range item.Copack {
+
+				done = timed(fmt.Sprintf("loading the co-packed mesh %s", cp.Filename))
+				coMesh, err := fauxgl.LoadMesh(cp.Filename)
+				if err != nil {
+					panic(err)
+				}
+				done()
+
+				// add coMesh to the main mesh. The "child"'s mesh is merged into its parent's.
+				mesh.Add(coMesh)
+			}
+
+			// 2. mesh centering.
+			done = timed("centering co-packed mesh")
+			mesh.Center()
+			done()
+
+			// 3. apply the scaling to the parent co-packing mesh (and implicitly its children).
+			//    Notice that if scaling is to be applied, it is done
+			//    before the computation of the BoundingBox and volume.
+			scale = item.Scale
+			scaleMatrix = fauxgl.Scale(fauxgl.V(scale, scale, scale))
+			if scale != 1.0 {
+				done = timed("scaling main co-packing mesh")
+				mesh.Transform(scaleMatrix)
+				done()
+			}
+
+			// 4. apply the manufacturing rotation to the parent co-packing mesh (and implicitly its children).
+			//    Notice that this is done before the computation of the BoundingBox and volume.
+			// IMPORTANT: do not confuse manufacturing orientation with the packing
+			//            orientations from the orientations provided by the annealing further on.
+			mfgRotationMatrix = getManufacturingOrientation(item)
+			mesh.Transform(mfgRotationMatrix)
+
+			// 5. update all the copies of the parent co-packing mesh
+			//    (and implicitly its children) for the json output.
+			size := mesh.BoundingBox().Size()
+			for i := 0; i < item.Count; i++ {
+				singleStlSize = append(singleStlSize, size)
+				srcStlNames = append(srcStlNames, item.Filename)
+				scaleStl = append(scaleStl, scaleMatrix)
+				mfgRotationStl = append(mfgRotationStl, mfgRotationMatrix)
+			}
+
+			fmt.Printf("  %d triangles\n", len(mesh.Triangles))
+			fmt.Printf("  %g x %g x %g\n", size.X, size.Y, size.Z)
+
+			// 6. coarse approx of its volume.
+			totalVolume += mesh.BoundingBox().Volume()
+		}
+
+		done = timed("building bvh tree")
+
+		model.Add(mesh, BVH_DETAIL, item.Count, spacing, getAvailableRotations(item.AxesLock))
+		ok = true
+		done()
+
+		fmt.Println("______________________________________________________")
+	}
+
+	if !ok {
+		fmt.Println("Usage: pack3d --input_config_json_filename==mesh_config.json --output_packing_json_filename=export.json")
+		fmt.Println(" - Packs N copies of each mesh into as small of a volume as possible.")
+		fmt.Println(" - Runs forever, looking for the best packing.")
+		fmt.Println(" - Results are written to disk whenever a new best is found.")
+		return
+	}
+
+	side := math.Pow(totalVolume, 1.0/3)
+	model.Deviation = side / 32 //it is not the distance between objects. And it seems that it will not reflect the distance.
+
+	/*  Mesh packing loop. This loop is to find the best STL mesh packing.
+	    Add 'break' in the loop to stop program */
+	start := time.Now()
+	maxItemNum := len(model.Items)
+	var timeLimit float64
+	fillVolumeWithSpacing := 0.0
+	totalFillVolume := 0.0
+	null := fauxgl.Matrix{
+		X00: 0, X01: 0, X02: 0, X03: 0,
+		X10: 0, X11: 0, X12: 0, X13: 0,
+		X20: 0, X21: 0, X22: 0, X23: 0,
+		X30: 0, X31: 0, X32: 0, X33: 0,
+	}
+	timeLimit = 10
+
+	minItemNum := 0
+	packItemNum := maxItemNum
+	successModel := NewModel()
+
+	for {
+		model, iterations = model.Pack(ANNEALING_ITERATIONS, nil, singleStlSize, frameSize, packItemNum)
+		/* iterations is the times of trial to find a output solution, if after trying for 100 times
+		   and no solution is found, then reset the model and try again. Usually if there is a solution,
+		   iterations will be 1 or 2 for most cases. */
+		if iterations >= 100 {
+			/* There is a case that even I reset the model for many times, I still can't find a solution,
+			   In this case, I need to set a threshold (20 second) to stop the software*/
+			if time.Since(start).Seconds() <= timeLimit {
+				model.Reset()
+				continue
+			} else {
+				// Linear search
+				//packItemNum -= 1
+
+				// Binary search
+				fmt.Println("Failed")
+				fmt.Println("packing#, max#, min# is: ", packItemNum, maxItemNum, minItemNum)
+				fmt.Println("-----------------------------------")
+				maxItemNum = packItemNum - 1
+				packItemNum = int(math.Ceil(float64((maxItemNum + minItemNum) / 2)))
+
+				model.Reset()
+				model.Transformation()[packItemNum] = null
+				start = time.Now()
+
+				if minItemNum > maxItemNum {
+					break
+				}
+
+				continue
+
+				//TODO: Unblock the following lines if want to return a json file including the error content
+				/*
+					err_content := err_msg{"Cannot get a result, please decrease your numbers of STLs or enlarge the frame sizes"}
+					fmt.Println(err_content.Error)
+					err_json, err := json.Marshal(err_content)
+					_, err := json.Marshal(err_content)
+					if err != nil{
+					fmt.Println("error:", err)
+					}
+					ioutil.WriteFile(fmt.Sprintf("%s.json", *fileNameArg), err_json, 0644)
+					break
+				*/
+			}
+		}
+
+		// Binary search
+		fmt.Println("Succeeded")
+		fmt.Println("packing#, max#, min# is: ", packItemNum, maxItemNum, minItemNum)
+		fmt.Println("-----------------------------------------")
+		minItemNum = packItemNum + 1
+		packItemNum = int(math.Ceil(float64((maxItemNum + minItemNum) / 2)))
+		successModel = model
+		start = time.Now()
+
+		if minItemNum > maxItemNum {
+			break
+		}
+		model.Reset()
+	}
+
+	done = timed("writing mesh")
+	var (
+		transMatrix    [4][4]float64
+		fillPercentage float64
+	)
+	transformation := successModel.Transformation()
+
+	// The scaling is applied directly in main.go and this is not ideal in terms of
+	// modularisation but for the sake of time it had to be squished in here.
+	// Tech debt: extract the scaling from main.go.
+	for j := 0; j < len(successModel.Items); j++ {
+		copack, ok := coPackMap[srcStlNames[j]]
+		if !ok {
+
+			t := transformation[j]
+			rt := t.Mul(mfgRotationStl[j]) // manufacturing rotation for the j-th mesh.
+			st := rt.Mul(scaleStl[j])      // scaled transformation for the j-th mesh.
+
+			fillVolumeWithSpacing = (singleStlSize[j].X + spacing) * (singleStlSize[j].Y + spacing) * (singleStlSize[j].Z + spacing)
+			if j < packItemNum {
+				totalFillVolume += fillVolumeWithSpacing
+				transMatrix = [4][4]float64{
+					{st.X00, st.X01, st.X02, st.X03},
+					{st.X10, st.X11, st.X12, st.X13},
+					{st.X20, st.X21, st.X22, st.X23},
+					{st.X30, st.X31, st.X32, st.X33},
+				}
+			} else {
+				transMatrix = [4][4]float64{
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+				}
+			}
+
+			// buildVolume's filling percentage.
+			fillPercentage = totalFillVolume / buildVolume
+
+			transMaps = append(transMaps, TransMap{srcStlNames[j], transMatrix, fillVolumeWithSpacing})
+
+		} else {
+
+			t := transformation[j]
+			rt := t.Mul(mfgRotationStl[j]) // manufacturing rotation for the j-th mesh.
+			st := rt.Mul(scaleStl[j])      // scaled transformation for the j-th mesh.
+			fillVolumeWithSpacing = (singleStlSize[j].X + spacing) * (singleStlSize[j].Y + spacing) * (singleStlSize[j].Z + spacing)
+			if j < packItemNum {
+				totalFillVolume += fillVolumeWithSpacing
+				transMatrix = [4][4]float64{
+					{st.X00, st.X01, st.X02, st.X03},
+					{st.X10, st.X11, st.X12, st.X13},
+					{st.X20, st.X21, st.X22, st.X23},
+					{st.X30, st.X31, st.X32, st.X33},
+				}
+			} else {
+				transMatrix = [4][4]float64{
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+					{0, 0, 0, 0},
+				}
+			}
+
+			// buildVolume's filling percentage.
+			fillPercentage = totalFillVolume / buildVolume
+
+			// Add the main co-packing mesh to transMaps.
+			transMaps = append(transMaps, TransMap{srcStlNames[j], transMatrix, fillVolumeWithSpacing})
+
+			// Add the co-packed meshes to transMaps.
+			for _, cp := range copack {
+				// IMPORTANT: The volume of a co-packed object is already included in the volume
+				//            of the parent object and the co-packed object's volume is set to 0.
+				transMaps = append(transMaps, TransMap{cp.Filename, transMatrix, 0})
+			}
+		}
+	}
+	positionsJson, err := json.Marshal(transMaps)
+	if err != nil {
+		fmt.Println("error:", err)
+	}
+	fmt.Println("the fill percentage is:", fillPercentage)
+	os.WriteFile(fmt.Sprintf("%s.json", output), positionsJson, 0644)
+	// os.Stdout.Write(positionsJson)
+
+	// STL file is no longer created, results returned as JSON for separate packer.
+	// Unblock on of the following lines to generate the packing STL file. This is typically done only for debugging purposes.
+	// model.Mesh().SaveSTL(fmt.Sprintf("pack3d_debug_test.stl")) // store the STL file in the main Nautilus folder.
+	// model.Mesh().SaveSTL(fmt.Sprintf("%s.stl", *fileNameArg))  // store the STL file next to the json file.
+	// model.TreeMesh().SaveSTL(fmt.Sprintf("out%dtree.stl", int(score*100000)))
+	done()
+}
+
+const (
+	BVH_DETAIL           = 8
+	ANNEALING_ITERATIONS = 2000000 // # of trials
+)
+
+/* This function returns the current time (it is a timer). */
+func timed(name string) func() {
+	if len(name) > 0 {
+		fmt.Printf("%s... ", name)
+	}
+	start := time.Now()
+	return func() {
+		fmt.Println(time.Since(start))
+	}
+}
+
+func getAvailableRotations(axesLock *AxesLock) []fauxgl.Matrix {
+	// This function returns only the available rotations
+	// which depend on the unlocked axes provided by the user.
+	// An unlocked axis is characterised by a `nil` theta angle.
+	// Setting a theta angle with a Float instead means that
+	// that rotation axis is locked to a specific angle.
+
+	// Tech debt: this function should probably be moved into model.go
+
+	if axesLock == nil {
+		return []fauxgl.Matrix{fauxgl.Identity()}
+	}
+	availableRotations := make([]fauxgl.Matrix, 0)
+	if axesLock.ThetaX == nil {
+		availableRotations = append(availableRotations, AxisXRotations...)
+	}
+	if axesLock.ThetaY == nil {
+		availableRotations = append(availableRotations, AxisYRotations...)
+	}
+	if axesLock.ThetaZ == nil {
+		availableRotations = append(availableRotations, AxisZRotations...)
+	}
+	// the function needs to return at least one dummy rotation (the identity matrix).
+	if len(availableRotations) == 0 {
+		availableRotations = append(availableRotations, fauxgl.Identity())
+	}
+	return availableRotations
+}
+
+func getManufacturingOrientation(item ConfigItem) fauxgl.Matrix {
+	// This function's purpose is to create a composite rotation matrix from the three provided angles.
+
+	// Tech debt: this function might need to be moved into a function in fauxgl.mesh.
+
+	// NOTE: The THREE.Euler's rotation order (in Rapidfab) has been set as 'ZYX' to match Blender's rotation order
+	//       and pack3d "seems" to be the same order of rotation but with the "minus" sign for all three angles.
+	//       e.g.: -fauxgl.Radians(*item.AxesLock.ThetaX)
+	mfgRotationMtx := fauxgl.Identity()
+	if item.AxesLock == nil {
+		return mfgRotationMtx
+	}
+	if item.AxesLock.ThetaX != nil {
+		axisX := AxisX.Vector() // x axis
+		mfgRotationMtx = mfgRotationMtx.Rotate(axisX, -fauxgl.Radians(*item.AxesLock.ThetaX))
+	}
+	if item.AxesLock.ThetaY != nil {
+		axisY := AxisY.Vector() // y axis
+		mfgRotationMtx = mfgRotationMtx.Rotate(axisY, -fauxgl.Radians(*item.AxesLock.ThetaY))
+	}
+	if item.AxesLock.ThetaZ != nil {
+		axisZ := AxisZ.Vector() // z axis
+		mfgRotationMtx = mfgRotationMtx.Rotate(axisZ, -fauxgl.Radians(*item.AxesLock.ThetaZ))
+	}
+	return mfgRotationMtx
+}
+
+type Config struct {
+	BuildVolume [3]float64   `json:"build_volume"`
+	Spacing     float64      `json:"spacing"`
+	ConfigItems []ConfigItem `json:"items"`
+}
+
+type ConfigItem struct {
+	Filename string    `json:"filename"`
+	Scale    float64   `json:"scale"`
+	Count    int       `json:"count"`
+	Copack   []*Copack `json:"copack,omitempty"`
+	AxesLock *AxesLock `json:"axes_lock"`
+}
+
+type Copack struct {
+	Filename string `json:"filename"`
+	// Scale        float64   `json:"scale"`
+	// Transformation [4][4]float64 `json:"transformation"`  // ch32838 initially required this field then the requirements changed.
+}
+
+// The struct name AxesLock is incorrect and should be replaced
+// with MfgOrientation and corrected everywhere else in this file.
+// This naming issue was spotted during the handoff to Tyler.
+type AxesLock struct {
+	ThetaX *float64 `json:"theta_x"`
+	ThetaY *float64 `json:"theta_y"`
+	ThetaZ *float64 `json:"theta_z"`
+}
